clean up code, delete action/reducers not needed
