+++ src/store/reducers/index.js
@@ -9,7 +9,6 @@ import { watchlist } from './watchlistReducer';
 import { movie } from './movieReducer';
 import { notwatchlist } from './notWatchListReducer';
 import { serviceProvider } from './serviceProviderReducer';
-import { removewatchlist } from './removeWatchListReducer';
 import { search } from "./searchReducer";
 import { landingPageReducer } from './landingPageReducer';
 
@@ -25,7 +24,6 @@ export const reducer = combineReducers({
   search,
   notwatchlist,
   serviceProvider,
-  removewatchlist,
   landingPageReducer
 });
